+++ src/Dropdown/Dropdown.tsx
@@ -22,6 +22,8 @@ export type DropdownProps<T> = {
     defaultButtonLabel?: string;
     items: DropdownItem<T>[];
     onSelect: (item: DropdownItem<T>) => void;
+    transparentButtonBg?: boolean;
+    minWidth?: boolean;
     disabled?: boolean;
     selectedItem: DropdownItem<T>;
     numItemsBeforeScroll?: number;
@@ -34,6 +36,8 @@ export default <T,>({
     defaultButtonLabel = '',
     items,
     onSelect,
+    transparentButtonBg = false,
+    minWidth = false,
     disabled = false,
     selectedItem,
     numItemsBeforeScroll = 0,
@@ -48,7 +52,11 @@ export default <T,>({
 
     return (
         <div
-            className={`tw-preflight tw-relative tw-w-full ${className}`}
+            className={classNames(
+                'tw-preflight tw-relative',
+                minWidth ? '' : 'tw-w-full',
+                className
+            )}
             onBlur={event => {
                 if (!event.currentTarget.contains(event.relatedTarget)) {
                     setIsActive(false);
@@ -61,7 +69,13 @@ export default <T,>({
             <button
                 id={id}
                 type="button"
-                className="tw-flex tw-h-8 tw-w-full tw-items-center tw-justify-between tw-border-0 tw-bg-gray-700 tw-px-2 tw-text-white"
+                className={classNames(
+                    'tw-flex tw-items-center tw-justify-between tw-border-0',
+                    minWidth ? '' : 'tw-w-full',
+                    transparentButtonBg
+                        ? 'tw-bg-transparent'
+                        : 'tw-h-8 tw-bg-gray-700 tw-px-2 tw-text-white'
+                )}
                 onClick={() => setIsActive(!isActive)}
                 disabled={disabled}
             >
@@ -71,7 +85,7 @@ export default <T,>({
                         : selectedItem.label}
                 </span>
                 <span
-                    className={`mdi mdi-chevron-down tw-text-lg ${classNames(
+                    className={`mdi mdi-chevron-down tw-text-lg/none ${classNames(
                         isActive && 'tw-rotate-180'
                     )}`}
                 />
@@ -93,8 +107,9 @@ export default <T,>({
                     numItemsBeforeScroll > 0 &&
                     items.length > numItemsBeforeScroll
                 }
-                className={`tw-text-while tw-absolute tw-right-0 tw-z-10 tw-w-full tw-border-t-2 tw-border-solid tw-border-gray-600 tw-bg-gray-700 tw-p-0 ${classNames(
+                className={`tw-text-while tw-absolute tw-z-10 tw-border-t-2 tw-border-solid tw-border-gray-600 tw-bg-gray-700 tw-p-0 ${classNames(
                     styles.content,
+                    minWidth ? '' : 'tw-right-0 tw-w-full',
                     !isActive && 'tw-hidden'
                 )}`}
             >